move decorator to local_test_server.py. Use decorator in real_call_test. Add tests for https & redirect.
