+++ tests/files_test.py
@@ -5,36 +5,13 @@
 """Post Files Tests"""
 # pylint: disable=line-too-long
 
-import functools
 import re
-import socketserver
-import threading
-import time
 from unittest import mock
 
 import mocket
 import pytest
 import requests as python_requests
-from local_test_server import LocalTestServerHandler
-
-
-def uses_local_server(func):
-    @functools.wraps(func)
-    def wrapper(*args, **kwargs):
-        with socketserver.TCPServer(("127.0.0.1", 5000), LocalTestServerHandler) as server:
-            server_thread = threading.Thread(target=server.serve_forever)
-            server_thread.daemon = True
-            server_thread.start()
-            time.sleep(2)  # Give the server some time to start
-
-            result = func(*args, **kwargs)
-
-            server.shutdown()
-            server.server_close()
-            time.sleep(2)
-            return result
-
-    return wrapper
+from local_test_server import uses_local_server
 
 
 @pytest.fixture
@@ -44,7 +21,6 @@ def log_stream():
 
 @pytest.fixture
 def post_url():
-    # return "https://httpbin.org/post"
     return "http://127.0.0.1:5000/post"
 
 
@@ -192,6 +168,7 @@ def test_post_files_file(  # pylint: disable=unused-argument
     assert sent.endswith(actual_request_post)
 
 
+@uses_local_server
 def test_post_files_complex(  # pylint: disable=unused-argument
     sock, requests, log_stream, post_url, request_logging
 ):
